chore (refs T22614): move DpEditableList to client/js.
