+++ client/js/components/core/DpEditableList.vue
