+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/basicSettings/DpEmailList.vue
@@ -40,7 +40,7 @@
 </template>
 
 <script>
-import DpEditableList from '@DemosPlanCoreBundle/components/DpEditableList'
+import DpEditableList from '@DpJs/components/core/DpEditableList'
 import { DpInput } from 'demosplan-ui/components'
 import validateEmail from '@DpJs/lib/validation/utils/validateEmail'
